step 3.2 and cleanup of other steps+LESSsampler + more
