+++ less/config/llama3_2/step1_warmup.yaml
@@ -36,8 +36,10 @@ save_adapter_weights_only: False
 dataset:
   _component_: torchtune.datasets.alpaca_cleaned_dataset
   packed: False  # True increases speed
-seed: null
+seed: 42
 shuffle: True
+same_shuffle_every_epoch: True
+n_print_examples: 3
 batch_size: 1 # LESS default = 1
 
 # Optimizer and Scheduler
